+++ src/main/java/com/pages/AccountsPage.java
@@ -3,18 +3,23 @@ package com.pages;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
 
 public class AccountsPage {
 
 	private WebDriver driver;
 
-	private By accountSections = By.cssSelector("div#center_column span");
+	//private By accountSections = By.cssSelector("div#center_column span");
+	
+	@FindBy(css = "div#center_column span")
+	List<WebElement> accountSections;
 
 	public AccountsPage(WebDriver driver) {
 		this.driver = driver;
+		PageFactory.initElements(driver,this);
 	}
 	
 	public String getAccountsPageTitle() {
@@ -22,15 +27,15 @@ public class AccountsPage {
 	}
 
 	public int getAccountsSectionCount() {
-		return driver.findElements(accountSections).size();
+		return accountSections.size();
 	}
 
 	public List<String> getAccountsSectionsList() {
 
 		List<String> accountsList = new ArrayList<>();
-		List<WebElement> accountsHeaderList = driver.findElements(accountSections);
+		//List<WebElement> accountsHeaderList = accountSections;
 
-		for (WebElement e : accountsHeaderList) {
+		for (WebElement e : accountSections) {
 			String text = e.getText();
 			System.out.println(text);
 			accountsList.add(text);
