+++ src/main/java/com/pages/ContactUsPage.java
@@ -1,22 +1,43 @@
 package com.pages;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.Select;
 
 public class ContactUsPage {
 	
 	private WebDriver driver;
-
+/*
 	private By subjectHeading = By.id("id_contact");
 	private By email = By.id("email");
 	private By orderRef = By.id("id_order");
 	private By messageText = By.id("message");
 	private By sendButton = By.id("submitMessage");
-	private By successMessg = By.cssSelector("div#center_column p");
-
+	private By successMessg = By.cssSelector("div#center_column p");*/
+	
+	@FindBy(id = "id_contact")
+	WebElement subjectHeading;
+	
+	@FindBy(id = "email")
+	WebElement email;
+	
+	@FindBy(id = "id_order")
+	WebElement orderRef;
+	
+	@FindBy(id = "message")
+	WebElement messageText;
+	
+	@FindBy(id = "id_contact")
+	WebElement sendButton;
+	
+	@FindBy(css = "div#center_column p")
+	WebElement successMessg;
+	
 	public ContactUsPage(WebDriver driver) {
 		this.driver = driver;
+		PageFactory.initElements(driver,this);
 	}
 
 	public String getContactUsPageTitle() {
@@ -24,19 +45,19 @@ public class ContactUsPage {
 	}
 	
 	public void fillContactUsForm(String heading, String emailId, String orderReference, String message) {
-		Select select = new Select(driver.findElement(subjectHeading));
+		Select select = new Select(subjectHeading);
 		select.selectByVisibleText(heading);
-		driver.findElement(email).sendKeys(emailId);
-		driver.findElement(orderRef).sendKeys(orderReference);
-		driver.findElement(messageText).sendKeys(message);
+		email.sendKeys(emailId);
+		orderRef.sendKeys(orderReference);
+		messageText.sendKeys(message);
 	}
 
 	public void clickSend() {
-		driver.findElement(sendButton).click();
+		sendButton.click();
 	}
 	
 	public String getSuccessMessg() {
-		return driver.findElement(successMessg).getText();
+		return successMessg.getText();
 	}
