+++ src/main/java/com/qa/factory/DriverFactory.java
@@ -2,16 +2,15 @@ package com.qa.factory;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.safari.SafariDriver;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 
 public class DriverFactory {
 
-	public WebDriver driver;
+	public static WebDriver driver;
 
-	public static ThreadLocal<WebDriver> tlDriver = new ThreadLocal<>();
+	//public static ThreadLocal<WebDriver> tlDriver = new ThreadLocal<>();
+	
 
 	/**
 	 * This method is used to initialize the thradlocal driver on the basis of given
@@ -20,34 +19,36 @@ public class DriverFactory {
 	 * @param browser
 	 * @return this will return tldriver.
 	 */
-	public WebDriver init_driver(String browser) {
+	public void init_driver(String browser) {
 
 		System.out.println("browser value is: " + browser);
 
 		if (browser.equals("chrome")) {
-			WebDriverManager.chromedriver().setup();
-			tlDriver.set(new ChromeDriver());
+			//WebDriverManager.chromedriver().setup();
+			//tlDriver.set(new ChromeDriver());
+			System.setProperty("webdriver.chrome.driver", "C:\\Users\\lilesh bansal\\git\\HybridFramework\\Drivers\\chromedriver.exe");
+		    driver = new ChromeDriver();
 		} else if (browser.equals("firefox")) {
 			WebDriverManager.firefoxdriver().setup();
-			tlDriver.set(new FirefoxDriver());
+			//tlDriver.set(new FirefoxDriver());
 		} else if (browser.equals("safari")) {
-			tlDriver.set(new SafariDriver());
+			//tlDriver.set(new SafariDriver());
 		} else {
 			System.out.println("Please pass the correct browser value: " + browser);
 		}
 
-		getDriver().manage().deleteAllCookies();
-		getDriver().manage().window().maximize();
-		return getDriver();
-
+		driver.manage().deleteAllCookies();
+		driver.manage().window().maximize();
+		//return driver;
+	}
 	}
 
 	/**
 	 * this is used to get the driver with ThreadLocal
 	 * 
 	 * @return
-	 */
+	 *//*
 	public static synchronized WebDriver getDriver() {
 		return tlDriver.get();
 	}
-}
+}*/
