+++ src/test/java/stepdefinitions/ApplicationHooks.java
@@ -1,0 +1,52 @@
+package stepdefinitions;
+
+import java.util.Properties;
+
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+
+import com.qa.factory.DriverFactory;
+import com.qa.util.ConfigReader;
+
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import io.cucumber.java.Scenario;
+
+public class ApplicationHooks extends DriverFactory{
+
+	//private DriverFactory driverFactory;
+	//public WebDriver driver;
+	private ConfigReader configReader;
+	Properties prop;
+
+	@Before(order = 0)
+	public void getProperty() {
+		configReader = new ConfigReader();
+		prop = configReader.init_prop();
+	}
+
+	@Before(order = 1)
+	public void launchBrowser() {
+		String browserName = prop.getProperty("browser");
+		//driverFactory = new DriverFactory();
+		init_driver(browserName);
+		
+	}
+
+	@After(order = 0)
+	public void quitBrowser() {
+		driver.quit();
+	}
+
+	@After(order = 1)
+	public void tearDown(Scenario scenario) {
+		if (scenario.isFailed()) {
+			// take screenshot:
+			String screenshotName = scenario.getName().replaceAll(" ", "_");
+			byte[] sourcePath = ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
+			scenario.attach(sourcePath, "image/png", screenshotName);
+
+		}
+	}
+
+}
