+++ src/test/java/testrunners/MyTestRunner.java
@@ -1,23 +1,18 @@
 package testrunners;
 
-import org.junit.runner.RunWith;
 
-import io.cucumber.junit.Cucumber;
-import io.cucumber.junit.CucumberOptions;
+import io.cucumber.testng.AbstractTestNGCucumberTests;
+import io.cucumber.testng.CucumberOptions;
 
-@RunWith(Cucumber.class)
+//@RunWith(Cucumber.class)
 @CucumberOptions(
-		features = {"src/test/resources/AppFeatures"},
-		glue = {"stepdefinitions", "AppHooks"},
-		plugin = {"pretty",
-				"com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:",
-				"timeline:test-output-thread/"
-
-				
-		}
+		features = {"src/test/resources/features"},
+		glue = {"stepdefinitions"},
+		plugin = {"pretty",	"com.aventstack.extentreports.cucumber.adapter.ExtentCucumberAdapter:",	"timeline:test-output-thread/"},
+		tags= "@contacttest"
 		
 		)
 
-public class MyTestRunner {
+public class MyTestRunner extends AbstractTestNGCucumberTests{
 
 }
